remove mydocuments code from backend
