--- backend/onyx/seeding/user_folders.yaml
@@ -1,6 +1,0 @@
-user_folders:
-  - id: -1
-    name: "Recent Documents"
-    description: "Documents uploaded by the user"
-    files: []
-    assistants: []
